Initial commit to create skeleton for the code-owners plugin
The code-owners plugin should provide support for defining code owners
for files in a repository.

If the code-owners plugin is enabled, changes will only be submittable
when all touched files are covered by approvals from code owners.

This plugin is an alternative to the existing find-owners plugin and the
owners plugin.

This initial commit only creates the skeleton for the code-owners
plugin, but doesn't implement any logic yet.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I00156c89ab50c4c0be9fc8c97389f68235b12eab