Explain that the plugin is specifically developed for Chrome/Android
While the plugin is kept as generic and configurable as possible, so
that it's suitable for everyone, it is specifically developed to satisfy
the requirements for code owners from the Chrome and Android teams at
Google. This means some of the functionality and design decisons are
driven by Google-specific use-cases.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I322d72663725094d89d05d392d05af7096c62592