Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Are you running something that detects missing Javadoc?  I've seen that the later versions of Error Prone have such checks.

No, I just noticed this by myself.

But last year we had a team effort to increase javadocs in Gerrit core [1] and for this Joerg Zieren created a report for us that listed classes with missing javadocs. But AFAIK this was created by a one-off script, not sure how it detected missing javadocs, could be that it was using Error Prone. If you are interested in this, we can ask Joerg. We would like the CI to point our missing javadoc, but before we can turn this on we would need to fix our current code base, which still seems to be a lot of work.

[1] https://gerrit-review.googlesource.com/q/hashtag:%22javadoc-dec19%22

Patch-set: 1