Define an abstraction for code owner configurations
Code owners are defined for folders in branches. Which syntax is used to
define the code owners and where the code owner configuration is stored
depends on the code owner backend.

With the CodeOwnerConfig class this change defines an abstraction for
code owner configurations that is independent of the syntax and the
storage location.

Since code owners are defined for folders in branches, the key of a
CodeOwnerConfig is a tuple of project + branch (BranchNameKey) and
folder.

For now we only support the simplest code owner configuration possible,
which is a set of code owners. Code owners in CodeOwnerConfigs are
represented as CodeOwnerReferences.

A CodeOwnerReference defines a code owner in a code owner
configuration. For now we only support to reference code owners by
email, but we may support further identifiers later, which is why we
have CodeOwnerReference as an abstraction instead of using Strings for
representing code owner emails in CodeOwnerConfig.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I7d2337fe2de9762f05ffbdab9165b1770eee4690