Support comment lines in find-owner code owner configuration files
Lines starting with '#' represent comments [1] and should be ignored.

According to [1] using '#' at the line beginning is the only possibility
to include comments (comments cannot appear at the end of a syntax line
and using '//' is not supported).

At the moment detecting comments doesn't make a big difference as they
would be detected as invalid lines and then be ignored. But later we
likely want to fail for invalid lines, but not for comments, hence it
makes sense to differentiate both cases. For now the only benefit is
that a warning will only be logged for invalid lines but not for comment
lines.

[1] https://gerrit.googlesource.com/plugins/find-owners/+/master/src/main/resources/Documentation/syntax.md

Change-Id: Ie99b79576076a02f7fd7ac6e3b933a6cf2d5f116
Signed-off-by: Edwin Kempin <ekempin@google.com>