Support writing CodeOwnerConfigs to find-owners syntax
This change adds a new 'format' method to CodeOwnerConfigParser that
allows to format a CodeOwnerConfig as string that represents the code
owner config in an 'OWNER' file. This method can be used to create code
owner configuration files programmatically, which we will need to do
during integration tests.

The tests in CodeOwnerConfigParserTest are rewritten so that they do:

1. parse a given code owner config
2. assert the parsed code owner config
3. format the parsed code owner config as string
4. assert the formatted string

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id9e48e680bc1eb37f7ad2e9e7edce97dca38381b