Add API for code owner backends
Having an abstraction for code owner backends allow us to support
different syntaxes and storages for code owner configurations.

For the start we have only a single method that allows us to get a code
owner config. Follow-up changes will add further methods (e.g. to
create/update code owner configs).

For testing the functionality we need to write a test code owner config
to a repository that we can then read through the new CodeOwnerBackend
API. This logic was already implemented in CodeOwnerConfigFileTest. To
avoid duplicating this logic, we extract it (and the logic to read code
owner configs directly from a repository) into a new FindOwnersTestUtil
class, which allows to reuse this logic from multiple tests.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id2799cf17efe9ef3695719730c5f9384d5539d84