Instantiate test API in base class to make it available for all tests
All acceptance tests will need to use the test API. Instead of
instantiating the test API in each of the test classes do it once in
AbstractCodeOwnersTest which is the base class for all acceptance tests.

Use a TestModule for running the tests so that we can bind the test API
in Guice, since otherwise it cannot be injected.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibcd45485449a2b15f880ebaae569f6f1656f25e1