Add enum for code owners backend IDs
Having an enum that allows us to iterate over all code owners backend
IDs will allow us to execute the integration tests once per code owners
backend (see follow-up change). An enum is needed because the test setup
is static and we cannot iterate over the registered code owners
backends, as they are only available after the plugin has been
installed.

To avoid that we forget to update the enum when a new code owners
backend is added, we have a test that checks that the enum matches the
registered code owners backends.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I13d17718b856ea4c3fd99b8dd48a51cf4c9e6359