Execute integrations tests for all code owner backends
Add a new base class for integration tests that defines a config suite
to run the tests once for every code owners backend.

The initialisation of the code owners API and the test API is moved
into this base class, since these APIs should only be used from
integration tests.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I249ac47facc42260fb0312ce595c2f912249dfd2