CodeOwnerConfig: Add method to compute local code owners for a path
The new method computes the local code owners for a path. Local code
owners are the code owners of the given path that are directly mentioned
in the code owner config (code owners from inherited and included code
owner configs are not considered).

At the moment the local code owners always match the code owners that
are returned from the codeOwners() method, but that's only for now while
we do not support per-file code onwers in code owner configs yet. Once
we add support for per-file code owners, we must check which of the
per-file expressions match the given path to know if the code owners of
the per-file expression  apply to this path. This matching logic will be
implemented by future changes in the new localCodeOwners() method.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Icac923dd2d194b56c034bd6b91b7ffaf2e947fa9