Add class that allows us to visit the owner configs that apply for a path
Code ownerships are inherited from code owner configs in parent folders.
This means to compute the code owners for a path we must inspect the
code owner configs of the paths folder and all its parent folders. To
make the computation of the effective code owners for a path easy add a
class that iterates over the parent folders and makes a callback for
each applying code owner config.

Later it will be possible to disable inheriting code owners from parent
folders, but since this is not implemented yet we always iterate through
the whole parent folder chain for now.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id79a29cda08ae4b080d495804db8b269b1dc3374