Add REST endpoint to list the code owners of a file in a branch
For the initial version of this REST endpoint we only implemented the
minimal functionality. It's just finding the code owners for a given
path and returns them. Non-resolveable code owners (emails that cannot
be resolved to an account or that are ambigious) are omitted from the
response.

For now we sort the results by account ID to have stable order for the
tests. This sort order is not documented because we intent to change it
soon (we intent to order the results by score but we don't have scores
yet).

For now the REST endpoint doesn't check the visibility of the returned
accounts. This will be implemented by a follow-up change.

Follow-up changes will also add options to the REST endpoint (e.g. to
request certain account options, to limit the result etc.).

For the sake of keeping the change small and reviewable this change also
doesn't add a Java API for listing code owners yet. This is also
something that will be added by a follow-up change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If7a21a7b906f056c29edd2b377a66b2f821c145d