Add Java API for GetCodeOwnersForPathInBranch REST endpoint
Adapt the acceptance test to use the Java API.

The method to query code owners in CodeOwners returns a QueryRequest
object rather than executing the query and returning the result
directly, so that we can support setting options on the query before
executing it. Request options will only be added by follow-up changes.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1f3d76b1f3c1e326cb7ab026542051659235aee9