GetCodeOwnersForPathInBranch: Sort code owners by distance score
Some code owners are considered better for reviewing/approving a path
than other code owners. Which code owners are better can depend on many
factors. One factor is the distance of the code owner configuration that
defines the code owner from the owned path. The smaller the distance the
better we consider the code owner as reviewer/approver for the path.

We represent the factors that contribute the to the decision which code
owner is considered better as CodeOwnerScore's. For now we consider only
the distance, but support for other scores may be added later.

Every code owner gets a scoring for each of the CodeOwnerScore's. The
scoring is a value between 0.0 and 1.0, where 0.0 is the lowest possible
scoring and 1.0 is the highest possible scoring. The scoring is computed
by deviding the scoring value by the max possible scoring value. If for
a score lower values are better (e.g. as for the distance score) the
scoring is inverted (1.0 - scoring).

CodeOwnerScoring is a class that collects and computes the scorings for
all code owners for a certain score.

The GetCodeOwnersForPathInBranch REST endpoint is adapted to sort code
owners by their scorings (code owners with higher scoring are returned
first). If code owners have the same scoring we still sort by account
ID, but we will change this in a follow-up change (hence this is not
documented in the REST API documentation).

For now the scoring of the code owners is not returned to the client.
That's the same that we do for the suggest reviewers REST endpoint in
Gerrit core. Also here the suggested reviewers are sorted by an
internal scoring but the scoring itself if not returned to the client.
It's likely that the frontend wants to know the scorings for the code
owners, but returning them to the client will then be implemented in a
separate change (e.g. at this point it's unclear whether the frontend
only needs the total scoring or details about the scorings on the
different scores).

The sorting only works if a scoring is available for all code owners and
fails if a scoring for a code owner is missing. This is because we
expect all code owners to be scored, and a missing scoring means that
something didn't work as intended.

Since we only have a single score for now (the distance score) we do not
implement any weighting yet. Weighting should be added as soon as we
have several scores, since not all scores are equally important.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I7de9788a7ce9687badcc63eb92c79db432dcbf0c