Return code owners in a random order if their scoring is the same
The GetCodeOwnersForPathInBranch will be used to suggest code owners for
files. Better suitable code owner already appear first in the list since
we sort the returned code owners by their scorings. If code owners had
the same scoring we returned them in a defined order (sorted by account
ID). This would have led to a disproportionate amount of reviews for
those code owner with lower account IDs since they always appear first
if the scoring is the same. To avoid this we now randomize the order in
which code owners with the same scoring are returned.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I2badf04e055738bfcd2f151727b53a9a0e255f6e