Add parser for reading/writing proto code owner configurations
Going forward we want to use a proto syntax for storing code owner
configs, as this syntax is easiler extensible, readable and maintainable
as the custom syntax that is used by the find-owners plugin. This change
starts adding a code owners backend that supports the proto syntax. With
this change only a parser for reading/writing proto code owner
configurations is added, other parts of the proto code owners backend
will be added by follow-up changes.

In contrast to the find-owners backend, the proto backend stores code
owner configs in OWNER_METADATA files (not OWNERS files).

The proto syntax that is added by this change only covers the minimum of
messages/fields that are necessary for defining a set of code owners for
a folder in a branch. The final proto syntax is not fully defined yet,
which means that the proto will still be changed. These changes will be
incompatible changes, which means the proto syntax should not be used in
any productive setups yet.

To avoid code duplication between code owners backends we define a
common interface for code owner config parsers that each backend can
implement for its syntax. The behaviour of the code owner config parsers
is mostly the same across backends. To avoid writing tests for the
common behaviour multiple times, we add an abstract base class for the
code owner config parser tests that covers the common logic.

Differences in behaviour exist in regards to handling of invalid lines,
invalid emails and inline comments (the proto backend fails parsing on
invalid lines, but accepts invalid emails and inline comments, the
find-owners backend ignores invalid lines, invalid emails and lines with
inline comments). These differences are OK for now and may be addressed
later (e.g. when we start implementing consistency checks). For now
(this change + follow-up changes) we are only interested in adding a
minimal working backend for the proto syntax.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6ed55133832fc12b29b95c53141ba2b2f41e5a63