Make CodeOwnerConfigFile generic so that it can be used by all backends
At the moment CodeOwnerConfigFile is part of the find-owners backend but
the same functionality is also needed in the proto backend. Currently
CodeOwnerConfigFile has 2 things that tie it to the find-owners backend,
the used file name and the used CodeOwnerConfigParser. To make
CodeOwnerConfigFile reusable we extract both (file name and
CodeOwnerConfigParser) and require callers to pass them in.

Alternatively I considered having a subclass for each backend, but since
there are only few callers which need to pass in the extra parameters,
passing in file name + CodeOwnerConfigParser seems less overhead than
maintaining a subclass per backend.

The check in CodeOwnerConfigFile to test whether the file exists had to
be slightly rewritten. This is because we are using the
CodeOwnerConfigParser interface now where the parse method throws an
IOException which prevents us from calling in from a lambda expression.

Since CodeOwnerConfigFile is generic now it is moved from the
backend.findowners package into the backend package.

Correspondingly also the test for this class was moved from the
backend.findowners package into the backend package. To be able to do
this the test was adapted to no longer use the
FindOwnersCodeOwnerConfigParser, but a backend independent test
implementation of CodeOwnerConfigParser (this test implementation is
just using Java serialization to format and parse CodeOwnerConfig's).

We also had FindOwnersUtil that depended on the find-owners specific
version of CodeOwnerConfigFile. This utility is made generic too (file
name and CodeOwnerConfigParser are passed in via a factory) and is now
renamed to BackendUtil. This way it can be used for tests across all
backends.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I2de3a4cbaee079b350c0fe8b9e569664df0e8cfb