Add a CodeOwnersBackend for the proto syntax
The implementation of the new backend is largely the same as the
existing FindOwnersBackend. The only difference is the name of the code
owner config files and the used CodeOwnerConfigParser. To avoid code
duplication implement an abstract backend that stores code owner configs
in files inside the corresponding folders in the source branch.

For testing CodeOwnersBackend implementations we now also have a base
class that implements the common tests, so that they do not need to be
duplicated.

The tests for deleting code owner configs are specific for the
FindOwnersBackend since the ProtoBackend doesn't delete empty code owner
configs. We may change this in a follow-up change to make the backends
behave more consistent.

All acceptance tests are automatically executed against the new backend
as well.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ie52cd9cca6330aa1698174652bf33a7ca660e16b