ProtoCodeOwnerConfigParser: Format empty code owner config as empty string
This causes a deletion of the code owner config file if the code owner
config is empty, rather than keeping the code owner config file with an
empty proto ("owners_config {\n}\n"). This makes the behaviour between
the ProtoCodeOwnerConfigParser and the FindOwnersCodeOwnerConfigParser
more consistent. As result also the tests get a little simpler.

Change-Id: Ide06ab239b464085c162426611185e082eec1502
Signed-off-by: Edwin Kempin <ekempin@google.com>