Copy code from find-owners plugin to reimplement FindOwnersCodeOwnerConfigParser
The FindOwnersBackend supports the syntax that is used by the
find-owners plugin. The find-owners plugin already has a parser for this
syntax and we do not want to reimplement it. So far we had a
self-written dummy implementation of the parser that enabled us to
implement the layers on top, but going forward we don't want to spend
time on developing our own parser, instead we just copy the parsing
logic from the find-owners plugin.

The parser in the find-owners plugin may not be implemented in an ideal
way (e.g. it uses regular expressions to match lines which is known to
be slow), but this is fine for us, since we only use this syntax until
the migration to the new syntax is done. Improving the parsing is not a
goal.

Using the original parsing code also ensures that we are as
backwards-compatible as possible.

As it turns out the find-owners plugin actually supports inline comments
in email lines. Adapt the corresponding test accordingly.

Change-Id: I4a55a513f4ef55a7f6093696ac08cfd1f2aae8b6
Signed-off-by: Edwin Kempin <ekempin@google.com>