FindOwnersBackend: Support ignoring parent code owners
Add a field to ignore parent code owners to CodeOwnerConfig and adapt
FindOwnersCodeOwnerConfigParser to populate this field when "set
noparent" is present in a code owner config file and to write the "set
noparent" line when a CodeOwnerConfig with ignoreParentCodeOwners() =
true is formatted.

Support for ignoring parent code owners in higher levels and in the
ProtoBackend will be added by following changes. This means for now
setting "set noparent" still has no effect for REST API calls.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I897666ded444d14dea7e46f13e63a3b1c2ea23ba