Add abstraction for a set of code owners
So far code owner configs can only contain a plain set of code owners
that own the complete folder in which the code owner config is stored.

Next we want to support defining code owners on path expressions. This
means code owner configs should allow to define sets of code owners that
own files matched by sets of path expressions.

In the code owners proto this will be expressed by code owner sets: A
code owner config can contain multiple code owner sets, each code owner
set contains a set of code owners and a set of path expressions for
which the ownership applies. Code owners for a complete folder are
expressed as a code owner set with an empty set of path expressions.

Since the code owners proto syntax will be our primary syntax we want to
keep the Java abstraction for code owner configs as close as possible
to it. For this reason we are adding a CodeOwnerSet abstraction that
defines a set of code owners for a set of path expressions. As a first
step (this change), we add the CodeOwnerSet class and use it to
represent the code owners for the whole folder. This means the
CodeOwnerSet contains the set of code owners and the set of path
expressions is empty aka doesn't exist yet (the field for the path
expressions will be added to CodeOwnerSet by a follow-up change).

Path expressions can only match files inside the folder (including
subfolders) in which the code owner config is stored. It will not be
possible to match files in other folders (parent folders or sibling
folders). Path expressions that match files in other folders will be
rejected as invalid on addition (a validation for this will be
implemented later).

The API's to read and write code owner configs are adapted to support
code owner sets. As a result of this we need to adapt all tests that
set/get code owners for a complete folder.

Modifying code owner sets requires a bit of boilerplate code. To avoid
that callers need to deal with it CodeOwnerSetModification offers a set
of static methods that do basic code owner set modifications.

Since there is no natural order for code owner sets, the code owner
config parsers/formatters keep the order of the code owner sets in the
code owner configs as it is, but duplicate code owner sets are filtered
out.

Change-Id: I52fc24a19ed2e51d7820d7a990410940f12df4f8
Signed-off-by: Edwin Kempin <ekempin@google.com>