CodeOwnerSet: Support path expressions
CodeOwnerSet has a new field for setting path expressions now. If path
expressions are set, the code owners that are defined in the
CodeOwnerSet only apply for paths that are matched by any of the path
expressions.

Which syntax is used for the path expressions (e.g. Java NIO globs,
regular expressions, Google3 path expressions) depends on the
CodeOwnersBackend. To not depend on any specific syntax for path
expressions we add PathExpressionMatcher as an abstraction for this that
must be implemented by the CodeOwnerBackends (the PathExpressionMatcher
checks for a given path expression if it matches a given path).

CodeOwnerBackends can choose to not support path expressions by not
allowing to use path expressions in the code owner config syntax. If the
code owner config files cannot contain path expressions, there are never
path expressions that need to be evaluated.

With this change none of the existing CodeOwnerBackends supports path
expressions yet (but follow-up changes will implement support for path
expressions in the different CodeOwnersBackends).

The pre-existing code assumed that we would implement the matching of
paths in CodeOwnerSet#matches(Path). However it turned out that this was
not a good choice since the logic that matches paths against path
expressions is specific to the CodeOwnersBackend that is used (this
means we would always need to pass the PathExpressionsMatcher into
CodeOwnerSet to evaluate path expressions). Instead we have now a
dedicated LocalCodeOwners class that computes local code owners for a
given path from a given CodeOwnerConfig. This class retrieves the
PathExpressionsMatcher from the CodeOwnersBackend and checks which of
the CodeOwnerSets in the CodeOwnerConfig match for the path.

Change-Id: I177d8088da9747c0025b4212ddd07acd066310c1
Signed-off-by: Edwin Kempin <ekempin@google.com>