Implement a PathExpressionMatcher for Google3 path expressions
This matcher will be used by the ProtoBackend to check if a path matches
a per-file path expression.

Since this path expression syntax is quite specific to Google, it's
unlikely that another code owner backend wants to use this path
expression syntax. Hence the Google3PathExpressionMatcher class is an
implementation detail of the proto backend and is not exposed to other
packages.

Google3 path expressions are similar to Java NIO globs, but less
powerful:

* '*': Represents any string that does not include slashes.
* '...': Represents any string, including slashes.
* Patterns such as '{**/,}' or '[1-4]' are not supported.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I875eac12713c69a1672bf06573d658cda85c20ba