Move code owner config test with non-sorted emails into test subclasses
At the moment the getCodeOwnerConfig method in
AbstractCodeOwnerConfigParserTest that constructs a code owner config
string is required to preserve the order of the passed in emails. This
is solely needed for the codeOwnerConfigWithNonSortedEmails() test and
fulfilling this requirement becomes a burden for adding tests for new
features. Drop this requirement for the getCodeOwnerConfig method and
construct the code owner config string for the
codeOwnerConfigWithNonSortedEmails() test manually. Since the manual
construction of the code owner config string is specific for the code
owner config syntax, move this test down into the specific test
subclasses.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibfe951a839138eddddae32d5cfba00730b850eac