Reject getting code owners for paths that don't exist in the change
The GetCodeOwnersForPathInChange REST endpoint should only return code
owners for files that exist in the change. To achieve this
CodeOwnersInChangeCollection rejects requests for paths that don't exist
in the change now.

For now, we reuse the existing logic in FileResource for this. In a
follow-up change we will rewrite this logic because we actually want to
allow requests for old paths of renamed and deleted paths which are
rejected now. Listing code owners for old paths of renamed and deleted
files should work because we will require code owner approval for the
old paths when a file is renamed and deleted, hence users should be able
to list the code owners for them. See TODO in
CodeOwnersInChangeCollection.

Since requests for paths that do not exist in the change are rejected
now, we must make sure for the integration tests that the test change is
created with all files that will be used in the tests. I considered
several approaches for this, but none was great. We now have a static
list of the file paths that will be used in the tests so that the test
setup knows which files it needs to create in the change. This is not
perfect, but the best approach I found. Since this is just test code
it's probably not worth spending much more time on making this nicer.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I34eb3dd5b3e914b49deb64e1d9fa313025a6c56f