Add REST endpoint to get code owners for files in a change
This REST endpoint basically has the same functionality as the existing
REST endpoint to get code owners for paths in a branch. A minor
difference will be that the REST endpoint to get code owners for files
in changes will later reject request for non-existing paths, but this
will only be implemented by a follow-up change.

The code owners for a file in a change are computed from the code owner
configuration files in the tip of the destination branch. This means
calling the REST endpoint to get code owners for a file in a change has
the same result as calling the REST endpoint to get code owners for the
file in the destination branch of the change. The REST endpoint to get
code owners for a file in a change only exists for convenience so that
callers do not need to know how the code owners for files in changes
are computed.

To avoid code duplication we add a base class for both REST endpoints
(AbstractGetCodeOwnersForPath) that contains the common logic.
Alternatively I considered letting the GetCodeOwnersForPathInChange REST
endpoint just delegate requests to GetCodeOwnersForPathInBranch REST
endpoint, but this would have had the disadvantage that the @Option
annotations would have been repeated in both REST endpoint classes and
GetCodeOwnersForPathInChange would have needed to take care to forward
all the options to GetCodeOwnersForPathInBranch.

Since GetCodeOwnersForPathInBranch and GetCodeOwnersForPathInChange
implement the same functionality, just under a different parent resource
(BranchResource vs. RevisionResource) the Java API is exactly the same
(CodeOwners) and we just have 2 different implementations of it
(CodeOwnersInBranchImpl and CodeOwnersInChangeImpl) that forward the
calls to the corresponding REST endpoints. The instantiation of the
CodeOwners API is done by CodeOwnersFactory. Depending on whether code
owners should be listed for a path in a branch or in a change it
instantiates CodeOwnersInBranchImpl or CodeOwnersInChangeImpl.

Since GetCodeOwnersForPathInBranch and GetCodeOwnersForPathInChange
implement the same functionality also the tests for both REST endpoints
have to cover the same functionality. Hence also here we have base
classes that implement the tests once, and then we cover both REST
endpoints via test subclasses.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I7efd57791b881ebd33622dcb2fe152188469625f