Extract logic to convert path to be compatible with Jgit into a class
The code-owners plugin uses absolute file paths (starting with a leading
'/'). The jgit API however doesn't accept leading '/' for absolute
paths, so we must remove the leading '/' if we pass an absolute path
into the jgit API. Since we will need this logic in several places,
extract it into an own class so that it can be reused easily.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I02c48d49deadd1d548cd1b5199080a85e024027f