Add class to compute the files which have been touched in a change
A new ChangedFiles class is added that can compute the files which have
been touched in a change. This functionality is needed to decide if code
owners can be retrieved for a file in a change (a follow-up change will
adapt CodeOwnersInChangeCollection to make use of ChangedFiles) or to
check if all files in a change are covered by code owner approvals when
a change is being submitted.

ChangedFiles computes the file diff newly on each access. Intentionally
we do not retrieve it from the PatchListCache which does a lot of
unneeded computations and hence would be slower. The Gerrit diff caches
are currently being redesigned. Once the envisioned ModifiedFilesCache
is available we should consider using it.

Instead of exposing DiffEntry's directly we wrap them in ChangedFile
objects which have better and smaller API.

ChangedFilesTest doesn't have an explicit test for renamed files. This
is because the test framework (PushOneCommit) doesn't have support to
add and delete files in the same change. Considering that detecting
renames is functionality of Gerrit core / JGit, we should be fine with
skipping this test here.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I8d74b9446aca3a845a1c0628dffd73f49f378b1c