Allow to get code owners for old path of renamed/deleted files in change
GetCodeOwnersForPathInChangeIT doesn't have an explicit test for getting
code owners for the old path of a renamed file. This is because the test
framework (PushOneCommit) doesn't have support to add and delete files
in the same change. However since this is very similar to getting code
owners for a deleted file, this should be fine for now.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id6dc7871b00b1ca54b924ad86fa62f4e7491a163