CodeOwnerResolver: Allow to skip the visibility check
Non-visible code owners should be filtered out when a user lists code
owners for a file. However when we check the submittability of a change
we want to take all code owners into account and hence disable the
visibility check when resolving code owners. The code owner approval
check for submit may only expose users that are reviewers of the change.
Since reviewers of a change are already visible to everyone who has
access to the change this will not expose users that the calling user
couldn't see otherwise.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I418eacc686af313a3ae5409008194935aa65d505