Allow to configure which approvals count as code owner approval
With the code-owners plugin every file in a change must be approved by a
code owner. Code owners approve modifications to files they own by
voting on a review label of the change. Which label this is and which
voting value counts as code owner approval should be configurable. By
default we require a +1 vote on the Code-Review label, but some projects
may prefer Code-Review+2 or a vote on a custom label (e.g.
Code-Owner-Review).

Which approval is required from code owners can be globally configured
in gerrit.config (parameter plugin.code-owners.requiredApproval).
Projects can override this by in the code-owners.config file in the
refs/meta/config branch (parameter codeOwners.requiredApproval).

The required approval must be set to a label and label value that do
exist. If a non-existing label or label value is set as required
approval the code owner configuration is invalid and trying to evaluate
code owner approvals will fail (a follow-up change may add a REST
endpoint to validate the code owner configuration so that issues can be
easily detected).

The configured required approval will be used for computing the code
owner statuses for the files in a change and for checking if a change is
submittable (will be implemented by follow-up changes).

Change-Id: I747217e89d9b85523b4ccef6f714718c5dfac779
Signed-off-by: Edwin Kempin <ekempin@google.com>