Add tests for renamed files
The PushOneCommit test API doesn't allow us to create changes with
renamed files. This is why tests for changes with renamed files have
been skipped so far. This change adds the missing tests. To create a
change with renamed files we use the change edit Java API now.

Add a method that creates a change with a renamed file to
AbstractCodeOwnersTest so that we need to implement this logic only
once. Since we use the change edit Java API we cannot return a
PushOneCommit.Result but only the change ID. Change the
createChangeWithFileDeletion method to also return the change ID instead
of PushOneCommit.Result to be consistent.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I4c7bfa3b5babe5a7ed5ef0eabadd6814d1b13dc8